Workingishidunnoi'mtired
Move picking seems to have issues not picking an empty space. I'm not sure why, and I literally can't stare at the screen more.

It absolutely is learning: starting from a blank brain file it goes from losing to taking advantage of oversights in the perfect trainer AI. I'm going to grab a couple hours of sleep and we can look at this before class.