AI works
The AI will minimax to pick the best move of all options from the entire decision tree.

I think. I hope.

Due to the small numnber of decisions to be made, the random draw can luck into the correct move a decent number of times, but nevertheless it does seem to work, and playing a bunch of repeat gamea bears this assessment out.