OthelloAI does alpha beta
OthelloAI now does alpha beta search. Backupai1 and 2.txt are included for legacy "not wanting to blow up previous files that might be used" purposes, but in all practicality they can be ignored. OthelloAI now does a proper alpha-beta search, with a depth pruning added.

That last is necessary for my machine since it sucks: to fully test it, turn the depth pruning up from 10 to like 15 or 20 and see if it works. Ideally, we want a score of 16 X 20 O, but I just could not have the program run in a reasonable time on this, so here you go.